finish installmentPlan

+++ src/main/kotlin/com/example/demo/dto/request/installmentPlan/InstallmentPlanDTO.kt
@@ -3,9 +3,9 @@ package com.example.demo.dto.request.installmentPlan
 import java.time.LocalDate
 
 data class InstallmentPlanDTO (
-    val cartItemId: Int = 0,
+    val cartItemId: Long = 0,
     val status: String = "PENDING",
-    val totalMonth: Double,
+    val totalMonth: Int,
     val startDate: LocalDate,
     val endDate: LocalDate,
     val installmentAmount: Double,
